feat(compression-evals): add PythonCode class in code.py
The PythonCode class is introduced in code.py. This class is used to
retrieve and handle Python code from any object. It includes methods
to get the raw code, type of the object, length of the code, and string
and repr representations of the object. A simple test is included in
the main function.