feat(compression): update return type and add transformers #compression-evals
This commit updates the return type of the `compress` method from `str` to `PythonCode` in the `compressors.py` file. It also adds a new transformer `RemoveEmptyLinesTransformer` to remove empty lines from a string in the `transformers.py` file.

In addition, the commit includes several changes to the `IdentifierRenamingTransformer` class in the `transformers.py` file. These changes include renaming variables and methods, updating the names in the store and load context, and transforming call and attribute nodes.

#NOTE at this time IdentifierRenamingTransformer is bugged (See Comment below docstring for info) and should not be used till fixed.